Factor out general document initialization into Application
